+++ client/utils/functions/toast.tsx
@@ -10,35 +10,8 @@ export const successToast = (message: string) => {
     pauseOnHover: true,
     draggable: true,
     progress: undefined,
+    className: "bg-base-green font-medium",
   });
-};
-
-export const errorHandler = (err?: AxiosError | any) => {
-  let errMessage: string = "Oops! Something went wrong.";
-  if (err) {
-    switch (err.response?.status) {
-      case 400:
-        errMessage = "Kindly check your inputs.";
-        break;
-      case 401:
-        errMessage = "Recaptcha Invalid. Try again later.";
-        break;
-      case 403:
-        errMessage = "Forbidden.";
-        break;
-      case 409:
-        errMessage = "Artifact already exists.";
-        break;
-      case 500:
-        errMessage = "Internal server error.";
-        break;
-      default:
-        errMessage = "Oops! Something went wrong.";
-        break;
-    }
-  }
-
-  errToast(errMessage);
 };
 
 export const errToast = (message: string) => {
@@ -50,5 +23,6 @@ export const errToast = (message: string) => {
     pauseOnHover: true,
     draggable: true,
     progress: undefined,
+    className: "font-medium bg-red-600",
   });
 };
